+++ lib/partition/uncoarsening/refinement/tabu_search/tabu_moves_queue.h
@@ -36,7 +36,7 @@ class tabu_moves_queue  {
                 bool empty();
 
                 void insert(NodeID node, PartitionID block, int time); 
-                int minValue();
+                int64_t minValue();
                 std::pair<NodeID, PartitionID> deleteMin();
 
                 bool contains(NodeID node, PartitionID block);
